BJS2-27102 Анализ комментариев
Исправлял конфликты
